Some fixes. Added some local app errors handling.

+++ widgets/mainwindow.cpp
@@ -96,37 +96,61 @@ void MainWindow::uploadAndRun()
 
 void MainWindow::runApp()
 {
+    bool isOkay = false;
     if (m_edisonCompileAct->isChecked()) {
         m_runAppAct->setDisabled(true);
         m_wifiConnection->runApp();
         m_runAppAct->setEnabled(true);
+        isOkay = true;
     }
     if (m_mingwCompileAct->isChecked()) {
         if (m_localApp->isOpen()) {
             m_localApp->kill();
         }
         QProcessEnvironment env = QProcessEnvironment::systemEnvironment();
-        QString pathToBins = QCoreApplication::applicationDirPath() + "mingw492_32\\murlibs";
-        env.insert("PATH", env.value("Path") + ";D:\\Projects\\RoboIDE\\build-RoboIDE-Desktop_Qt_5_5_1_MinGW_32bit-Debug\\debug\\mingw492_32\\murlibs");
+        env.insert("PATH", SettingsManager::instance().mingwBinarysPath());
         m_localApp->setProcessEnvironment(env);
         m_localApp->start(m_sourceCompiller->pathToBinary());
-        qDebug() << m_localApp->environment();
 
+        if (m_localApp->state() != QProcess::NotRunning) {
+            isOkay = true;
+        }
     }
-    m_roboIdeConsole->append("Программа запущена!");
+
+    if (isOkay) {
+        m_roboIdeConsole->append("Программа запущена!");
+    }
+    else {
+        m_roboIdeConsole->append("Во время запуска программы произошла ошибка.");
+        if (m_mingwCompileAct->isChecked()) {
+            m_roboIdeConsole->append(m_localApp->errorString());
+        }
+    }
 }
 
 void MainWindow::killApp()
 {
+    bool isOk = false;
     if (m_edisonCompileAct->isChecked()) {
         m_stopAppAct->setDisabled(true);
         m_wifiConnection->killApp();
         m_stopAppAct->setEnabled(true);
     }
     if (m_mingwCompileAct->isChecked()) {
-        m_localApp->kill();
+        if (m_localApp->isOpen()) {
+            m_localApp->kill();
+            isOk = true;
+        }
+
     }
-    m_roboIdeConsole->append("Программа остановлена!");
+    if (isOk) {
+        m_roboIdeConsole->append("Программа остановлена!");
+    }
+    else {
+        if (m_mingwCompileAct->isChecked()) {
+            m_roboIdeConsole->append("Программа не была запущена. Нечего остонавливать.");
+        }
+    }
 }
 
 void MainWindow::openFileOrProject()
@@ -158,11 +182,13 @@ void MainWindow::openHelp()
 void MainWindow::switchCompilationTargetToEdison()
 {
     m_mingwCompileAct->setChecked(false);
+    m_edisonCompileAct->setChecked(true);
 }
 
 void MainWindow::switchCompilationTargetToDesktop()
 {
     m_edisonCompileAct->setChecked(false);
+    m_mingwCompileAct->setChecked(true);
 }
 
 void MainWindow::processOutReceived()
@@ -187,11 +213,9 @@ void MainWindow::createActions()
 {
     //!Edit file actions
 
-    m_redoAct = new QAction(QIcon(":/icons/icons/tools/try-me-icon.png"), tr("Повторить"), this);
+    m_redoAct = new QAction(QIcon(":/icons/icons/tools/redo.png"), tr("Повторить"), this);
     m_redoAct->setShortcut(QKeySequence::Redo);
     m_redoAct->setIconVisibleInMenu(false);
-    m_redoAct->setDisabled(true);
-
 
     m_undoAct = new QAction(QIcon(":/icons/icons/tools/undo.png"), tr("Отменить"), this);
     m_undoAct->setShortcut(QKeySequence::Undo);
@@ -248,10 +272,10 @@ void MainWindow::createActions()
     m_openHelpAct = new QAction(tr("Справка"), this);
     m_openHelpAct->setShortcut(QKeySequence(Qt::Key_F1));
 
-    m_edisonCompileAct = new QAction(tr("MUR Robot"), this);
+    m_edisonCompileAct = new QAction(tr("Робот"), this);
     m_edisonCompileAct->setCheckable(true);
     m_edisonCompileAct->setChecked(true);
-    m_mingwCompileAct = new QAction(tr("Desktop"), this);
+    m_mingwCompileAct = new QAction(tr("Компьютер"), this);
     m_mingwCompileAct->setCheckable(true);
 }
